rag_project/backend/requirements.txt
fastapi
uvicorn
transformers
faiss-cpu
sentence-transformers
